+++ rust/tauri-app/src/boot/timeago.js
@@ -1,5 +1,0 @@
-import timeago from "vue-timeago3"
-
-export function setupTimeago (app) {
-  app.use(timeago)
-}