+++ src/walletConnect/provider.js
@@ -1,4 +1,0 @@
-import WalletConnectProvider from "@walletconnect/web3-provider"
-export const provider = new WalletConnectProvider({
-  infuraId: "f973532779114df7a8bc4ba8b9977cfb"
-})